add repository

+++ app/src/main/java/com/istea/appdelclima/presentacion/clima/ClimaEstado.kt
@@ -3,9 +3,9 @@ package com.istea.appdelclima.presentacion.clima
 sealed class ClimaEstado {
     data class Exitoso (
         val ciudad: String = "",
-        val temperatura: Int = 0,
+        val temperatura: Double = 0.0,
         val descripcion: String= "",
-        val st :Int = 0,
+        val st :Double = 0.0,
     ) : ClimaEstado()
     data class Error(
         val mensaje :String = "",
